// src/app/shared/components/enums/status-sinapse-enum.ts
export enum StatusSinapseEnum {
    // OUTROS = 0,
    // PENDENTE = 1,
    // CONFIRMADO = 2,
    // CANCELADO = 3,
    // FINALIZADO = 4,
    // PEN_IMP = 5,
    // CONF_IMP = 6,
    // ERRO = 7

    Pendente = 1,
    Confirmada = 2,
    Cancelada = 3,
    Finalizada = 4,
    PendenteImpedida = 5,
    ConfirmadaImpedida = 6,
    Erro = 7,
    CienciaInformada = 8,
}